Refactor: Move restaurant tests from main.rs to restaurants.rs
Moved all tests related to restaurant functionality from the main.rs module to the restaurants.rs module.
This includes the helper function init_test_db_app and all associated test cases.

Updated imports and cleaned up main.rs accordingly. All tests pass.
